Enhance tests in `test_simple_gla.py`
